requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
websockets>=12.0
httpx>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0
sarvamai>=0.1.27

# ── RAG / Vector Store ────────────────────────────────────────────────────────
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
chromadb>=0.4.24
sentence-transformers>=2.7.0

# ── Document Loaders ──────────────────────────────────────────────────────────
pymupdf>=1.24.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=5.0.0

# ── Performance ───────────────────────────────────────────────────────
          # 2x faster async event loop (optional but recommended)

# ── Utilities ─────────────────────────────────────────────────────────
aiofiles>=23.2.0
numpy>=1.26.0

# ── OpenAI (GPT-4o-mini LLM) ─────────────────────────────────────────
openai>=1.30.0

# ── LiveKit + Vobiz Telephony ─────────────────────────────────────────
livekit>=1.1.0
livekit-agents>=1.4.0
livekit-api>=1.1.0
livekit-plugins-openai>=1.4.0
livekit-plugins-sarvam>=1.4.0
livekit-plugins-silero>=1.4.0
livekit-plugins-deepgram>=1.4.0

